+++ postgres/database-dump/main.tf
@@ -1,4 +1,5 @@
 data "aws_caller_identity" "current" {}
+data "aws_region" "current" {}
 
 data "aws_iam_policy_document" "allow_task_creation" {
   statement {
@@ -22,8 +23,8 @@ data "aws_iam_policy_document" "allow_task_creation" {
       "logs:PutLogEvents"
     ]
     resources = [
-      "arn:aws:logs:eu-west-2:${data.aws_caller_identity.current.account_id}:log-group:/ecs/${local.task_name}",
-      "arn:aws:logs:eu-west-2:${data.aws_caller_identity.current.account_id}:log-group:/ecs/${local.task_name}:log-stream:*",
+      "arn:aws:logs:${local.region_account}:log-group:/ecs/${local.task_name}",
+      "arn:aws:logs:${local.region_account}:log-group:/ecs/${local.task_name}:log-stream:*",
     ]
   }
 }
@@ -40,6 +41,21 @@ data "aws_iam_policy_document" "assume_ecs_task_role" {
     }
 
     actions = ["sts:AssumeRole"]
+  }
+
+  dynamic "statement" {
+    for_each = local.pipeline_tasks
+    content {
+      sid    = "AllowPipelineAssumeRole"
+      effect = "Allow"
+
+      principals {
+        type        = "AWS"
+        identifiers = ["arn:aws:iam::${coalesce(statement.value.to_account, data.aws_caller_identity.current.account_id)}:role/${var.database_name}-${statement.value.from}-to-${statement.value.to}-copy-pipeline-codebuild"]
+      }
+
+      actions = ["sts:AssumeRole"]
+    }
   }
 }
 
@@ -104,6 +120,132 @@ resource "aws_iam_role_policy" "allow_data_dump" {
   name   = "AllowDataDump"
   role   = aws_iam_role.data_dump.name
   policy = data.aws_iam_policy_document.data_dump.json
+}
+
+resource "aws_iam_role_policy" "allow_pipeline_access" {
+  for_each = toset(length(local.pipeline_tasks) > 0 ? [""] : [])
+  name     = "AllowPipelineAccess"
+  role     = aws_iam_role.data_dump.name
+  policy   = data.aws_iam_policy_document.pipeline_access.json
+}
+
+data "aws_iam_policy_document" "pipeline_access" {
+  policy_id = "pipeline_access"
+  statement {
+    sid    = "AllowListAccountAliases"
+    effect = "Allow"
+    actions = [
+      "iam:ListAccountAliases",
+    ]
+    resources = [
+      "*",
+    ]
+  }
+
+  statement {
+    sid    = "AllowGetCopilotMetaData"
+    effect = "Allow"
+    actions = [
+      "ssm:GetParametersByPath",
+      "ssm:GetParameters",
+      "ssm:GetParameter"
+    ]
+    resources = [
+      "arn:aws:ssm:${local.region_account}:parameter/copilot/*",
+    ]
+  }
+
+  statement {
+    sid    = "AllowReadOnRDSSecrets"
+    effect = "Allow"
+    actions = [
+      "secretsmanager:GetSecretValue",
+    ]
+    resources = [
+      "arn:aws:secretsmanager:${local.region_account}:secret:rds*"
+    ]
+  }
+
+  statement {
+    sid    = "AllowRunningDumpTask"
+    effect = "Allow"
+    actions = [
+      "ecs:RunTask",
+    ]
+    resources = [
+      "arn:aws:ecs:${local.region_account}:task-definition/*-dump:*"
+    ]
+  }
+
+  statement {
+    sid    = "AllowLogTrail"
+    effect = "Allow"
+    actions = [
+      "logs:StartLiveTail",
+    ]
+    resources = [
+      "arn:aws:logs:${local.region_account}:log-group:/ecs/*-dump"
+    ]
+  }
+
+  statement {
+    sid    = "AllowPassRoleToTaskExec"
+    effect = "Allow"
+    actions = [
+      "iam:PassRole",
+    ]
+    resources = [
+      "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/*-dump-exec"
+    ]
+  }
+
+  statement {
+    sid    = "AllowDescribeLogs"
+    effect = "Allow"
+    actions = [
+      "logs:DescribeLogGroups",
+    ]
+    resources = [
+      "arn:aws:logs:${local.region_account}:log-group::log-stream:"
+    ]
+  }
+
+  statement {
+    sid = "AllowRedisListVersions"
+    actions = [
+      "elasticache:DescribeCacheEngineVersions"
+    ]
+    effect = "Allow"
+    resources = [
+      "*"
+    ]
+  }
+
+  statement {
+    sid = "AllowOpensearchListVersions"
+    actions = [
+      "es:ListVersions",
+      "es:ListElasticsearchVersions"
+    ]
+    effect = "Allow"
+    resources = [
+      "*"
+    ]
+  }
+
+  statement {
+    sid    = "AllowDescribeVPCsAndSubnets"
+    effect = "Allow"
+    actions = [
+      "ec2:DescribeVpcs",
+      "ec2:DescribeSubnets",
+      "ec2:DescribeRouteTables",
+      "ec2:DescribeSecurityGroups"
+    ]
+    resources = [
+      "*"
+    ]
+  }
 }
 
 resource "aws_ecs_task_definition" "service" {