+++ postgres/database-load/locals.tf
@@ -14,5 +14,8 @@ locals {
   dump_kms_key_alias = "alias/${local.dump_task_name}"
   dump_bucket_name   = local.dump_task_name
 
+  pipeline_task  = lookup(var.task, "pipeline", null) != null
+  region_account = "${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}"
+
   ecr_repository_arn = "arn:aws:ecr-public::763451185160:repository/database-copy"
 }